fix scalahostNsc tests
Since Database no longer has .toBinary method, I've temporarily disabled
serialization roundtrip checks until we get to #827.
