+++ scalahost/nsc/src/test/scala/scala/meta/tests/scalahost/mirrors/DatabaseSuite.scala
@@ -1,241 +1,192 @@
 package scala.meta.tests
 package scalahost
 
-import scala.meta._
+import org.scalatest._
+import java.io.{File, PrintWriter}
+import scala.collection.immutable.Seq
+import scala.reflect.io._
+import scala.tools.cmd.CommandLineParser
+import scala.tools.nsc.{CompilerCommand, Global, Settings}
+import scala.tools.nsc.reporters.StoreReporter
+import scala.compat.Platform.EOL
+import scala.{meta => m}
+import scala.meta.internal.semantic.DatabaseOps
 
-class DatabaseSuite extends OnlineMirrorSuite {
-  names(
-    """
-    |object First {
-    |  def main(args: Array[String]): Unit = {
-    |    val list = List(1, 2, 3)
-    |    println(list)
-    |  }
-    |}
-  """.trim.stripMargin,
-    """
-    |[7..12): First => _empty_.First.
-    |[21..25): main => _empty_.First.main([Ljava/lang/String;)V.
-    |[26..30): args => _empty_.First.main([Ljava/lang/String;)V.(args)
-    |[32..37): Array => _root_.scala.Array#
-    |[38..44): String => _root_.scala.Predef.String#
-    |[48..52): Unit => _root_.scala.Unit#
-    |[65..69): list => <...>@61..85
-    |[72..76): List => _root_.scala.collection.immutable.List.apply(Lscala/collection/Seq;)Lscala/collection/immutable/List;.
-    |[90..97): println => _root_.scala.Predef.println(Ljava/lang/Object;)V.
-    |[98..102): list => <...>@61..85
-  """.trim.stripMargin
-  )
+abstract class DatabaseSuite extends FunSuite { self =>
+  private def test(code: String)(fn: => Unit): Unit = {
+    var name = code.trim.replace(EOL, " ")
+    if (name.length > 50) name = name.take(50) + "..."
+    super.test(name)(fn)
+  }
 
-  targeted(
-    """
-    |object <<Second>> {
-    |  def doSomething = {
-    |    42
-    |  }
-    |}
-  """.trim.stripMargin,
-    second => {
-      assert(second.symbol === Symbol("_empty_.Second."))
+  lazy val g: Global = {
+    def fail(msg: String) = sys.error(s"DatabaseSuite initialization failed: $msg")
+    val classpath = System.getProperty("sbt.paths.scalahost.test.classes")
+    val pluginpath = System.getProperty("sbt.paths.scalahost.compile.jar")
+    val options = "-Yrangepos -Ywarn-unused-import -cp " + classpath + " -Xplugin:" + pluginpath + ":" + classpath + " -Xplugin-require:scalahost"
+    val args = CommandLineParser.tokenize(options)
+    val emptySettings = new Settings(error => fail(s"couldn't apply settings because $error"))
+    val reporter = new StoreReporter()
+    val command = new CompilerCommand(args, emptySettings)
+    val settings = command.settings
+    val g = new Global(settings, reporter)
+    val run = new g.Run
+    g.phase = run.parserPhase
+    g.globalPhase = run.parserPhase
+    g
+  }
+  private lazy val databaseOps: DatabaseOps { val global: self.g.type } = new DatabaseOps {
+    val global: self.g.type = self.g
+  }
+  import databaseOps._
+
+  private def computeDatabaseFromSnippet(code: String): m.Database = {
+    val cwd = sys.props("user.dir")
+    try {
+      val javaFile = File.createTempFile("paradise", ".scala")
+      val writer = new PrintWriter(javaFile)
+      try writer.write(code)
+      finally writer.close()
+
+      sys.props("user.dir") = javaFile.getParentFile.getAbsolutePath
+      val run = new g.Run
+      val abstractFile = AbstractFile.getFile(javaFile)
+      val sourceFile = g.getSourceFile(abstractFile)
+      val unit = new g.CompilationUnit(sourceFile)
+      run.compileUnits(List(unit), run.phaseNamed("terminal"))
+
+      g.phase = run.parserPhase
+      g.globalPhase = run.parserPhase
+      val reporter = new StoreReporter()
+      g.reporter = reporter
+      unit.body = g.newUnitParser(unit).parse()
+      val errors = reporter.infos.filter(_.severity == reporter.ERROR)
+      errors.foreach(error => fail(s"scalac parse error: ${error.msg} at ${error.pos}"))
+
+      val packageobjectsPhase = run.phaseNamed("packageobjects")
+      val phases = List(run.parserPhase, run.namerPhase, packageobjectsPhase, run.typerPhase)
+      reporter.reset()
+
+      phases.foreach(phase => {
+        g.phase = phase
+        g.globalPhase = phase
+        phase.asInstanceOf[g.GlobalPhase].apply(unit)
+        val errors = reporter.infos.filter(_.severity == reporter.ERROR)
+        errors.foreach(error => fail(s"scalac ${phase.name} error: ${error.msg} at ${error.pos}"))
+      })
+      g.phase = run.phaseNamed("patmat")
+      g.globalPhase = run.phaseNamed("patmat")
+
+      m.Database(List(unit.source.toInput -> unit.toAttributes))
+    } finally {
+      sys.props("user.dir") = cwd
     }
-  )
+  }
 
-  names(
-    """
-    |import _root_.scala.List
-    |
-    |class C {
-    |  _root_.scala.List
-    |}
-  """.trim.stripMargin,
-    """
-    |[7..13): _root_ => _root_.
-    |[14..19): scala => _root_.scala.
-    |[20..24): List => _root_.scala.package.List.;_root_.scala.package.List#
-    |[32..33): C => _empty_.C#
-    |[38..44): _root_ => _root_.
-    |[45..50): scala => _root_.scala.
-    |[51..55): List => _root_.scala.collection.immutable.
-  """.trim.stripMargin
-  )
+  private def computeDatabaseSectionFromSnippet(code: String, sectionName: String): String = {
+    val database = computeDatabaseFromSnippet(code)
+    val path = g.currentRun.units.toList.last.source.file.file.getAbsolutePath
+    val payload = database.toString.split(EOL)
+    val section = payload.dropWhile(_ != sectionName + ":").drop(1).takeWhile(_ != "")
+    section.mkString(EOL).replace(path, "<...>")
+  }
 
-  targeted(
-    // curried function application with named args, #648
-    """
-      |object Foo {
-      |  def bar(children: Int)(x: Int) = children + x
-      |  <<bar>>(children = 4)(3)
-      |}
-    """.trim.stripMargin,
-    second => {
-      assert(second.symbol === Symbol("_empty_.Foo.bar(II)I."))
+  def names(code: String, expected: String): Unit = {
+    test(code)(assert(expected === computeDatabaseSectionFromSnippet(code, "Names")))
+  }
+
+  def messages(code: String, expected: String): Unit = {
+    test(code)(assert(expected === computeDatabaseSectionFromSnippet(code, "Messages")))
+  }
+
+  def denotations(code: String, expected: String): Unit = {
+    test(code)(assert(expected === computeDatabaseSectionFromSnippet(code, "Denotations")))
+  }
+
+  def sugars(code: String, expected: String): Unit = {
+    test(code)(assert(expected === computeDatabaseSectionFromSnippet(code, "Sugars")))
+  }
+
+  private def computeDatabaseAndNamesFromMarkup(markup: String): (m.Database, List[m.Name]) = {
+    val chevrons = "<<(.*?)>>".r
+    val ps0 = chevrons.findAllIn(markup).matchData.map(m => (m.start, m.end)).toList
+    val ps = ps0.zipWithIndex.map { case ((s, e), i) => (s - 4 * i, e - 4 * i - 4) }
+    val code = chevrons.replaceAllIn(markup, "$1")
+    val database = computeDatabaseFromSnippet(code)
+    val unit = g.currentRun.units.toList.last
+    val source = unit.toSource
+    val names = ps.map {
+      case (s, e) =>
+        val names = source.collect {
+          case name: m.Name if name.pos.start.offset == s && name.pos.end.offset == e => name
+        }
+        val chevron = "<<" + code.substring(s, e) + ">>"
+        names match {
+          case Nil => sys.error(chevron + " does not wrap a name")
+          case List(name) => name
+          case _ => sys.error("fatal error processing " + chevron)
+        }
     }
-  )
+    (database, names)
+  }
 
-  targeted(
-    """
-      |case class User(name: String, age: Int)
-      |object Library {
-      |  val u: User = ???
-      |  u.<<copy>>(<<age>> = 43)
-      |}
-    """.trim.stripMargin,
-    (copy, age) => {
-      assert(copy.symbol === Symbol("_empty_.User#copy(Ljava/lang/String;I)LUser;."))
-      assert(age.symbol === Symbol("_empty_.User#copy(Ljava/lang/String;I)LUser;.(age)"))
+  trait OverloadHack1; implicit object OverloadHack1 extends OverloadHack1
+  trait OverloadHack2; implicit object OverloadHack2 extends OverloadHack2
+  trait OverloadHack3; implicit object OverloadHack3 extends OverloadHack3
+  trait OverloadHack4; implicit object OverloadHack4 extends OverloadHack4
+  trait OverloadHack5; implicit object OverloadHack5 extends OverloadHack5
+
+  def targeted(markup: String, fn: m.Database => () => Unit)(implicit hack: OverloadHack1): Unit = {
+    test(markup) {
+      val (database, names) = computeDatabaseAndNamesFromMarkup(markup)
+      names match {
+        case List() => fn(database)()
+        case _ => sys.error(s"0 chevrons expected, ${names.length} chevrons found")
+      }
     }
-  )
+  }
 
-  denotations(
-    """
-      |import scala.language.experimental.macros
-      |
-      |package foo {
-      |  class C1(p1: Int, val p2: Int, var p3: Int) {
-      |    def this() = this(0, 0, 0)
-      |    val f1 = {
-      |      val l1 = ???
-      |      var l2 = ???
-      |      ???
-      |    }
-      |    var f2 = ???
-      |    def m1[T](x: Int): Int = ???
-      |    def m2 = macro ???
-      |    type T1 <: Int
-      |    type T2 = Int
-      |  }
-      |
-      |  abstract class C2 {
-      |    def m3: Int
-      |    final def m4 = ???
-      |  }
-      |
-      |  sealed class C3 extends C2 {
-      |    def m3 = 42
-      |    override def toString = ""
-      |  }
-      |
-      |  trait T {
-      |    private val f1 = ???
-      |    private[this] val f2 = ???
-      |    private[foo] val f3 = ???
-      |    protected var f4 = ???
-      |    protected[this] var f5 = ???
-      |    protected[foo] var f6 = ???
-      |  }
-      |
-      |  object M {
-      |    implicit def i1 = ???
-      |    lazy val l1 = ???
-      |    case class C1()
-      |    class C2[+T, -U]
-      |  }
-      |}
-      |
-      |package object foo {
-      |}
-  """.trim.stripMargin,
-    """
-      |_root_.foo. => package foo
-      |_root_.foo.C1# => class C1
-      |_root_.foo.C1#(p1) => param p1: Int
-      |_root_.foo.C1#(p2) => val param p2: Int
-      |_root_.foo.C1#(p3) => var param p3_=: (x$1: Int)Unit
-      |_root_.foo.C1#T1# => abstract type T1:  <: Int
-      |_root_.foo.C1#T2# => type T2: Int
-      |_root_.foo.C1#`<init>`()V. => secondaryctor <init>: ()foo.C1
-      |_root_.foo.C1#`<init>`(III)V. => primaryctor <init>: (p1: Int, p2: Int, p3: Int)foo.C1
-      |_root_.foo.C1#f1. => val f1: Nothing
-      |_root_.foo.C1#f1.l1. => val l1: Nothing
-      |_root_.foo.C1#f1.l2. => var l2: Nothing
-      |_root_.foo.C1#f2. => var f2_=: (x$1: Nothing)Unit
-      |_root_.foo.C1#m1(I)I. => def m1: [T](x: Int)Int
-      |_root_.foo.C1#m1(I)I.(x) => param x: Int
-      |_root_.foo.C1#m1(I)I.T# => typeparam T
-      |_root_.foo.C1#m2()Lscala/Nothing;. => macro m2: Nothing
-      |_root_.foo.C2# => abstract class C2
-      |_root_.foo.C2#`<init>`()V. => primaryctor <init>: ()foo.C2
-      |_root_.foo.C2#m3()I. => abstract def m3: Int
-      |_root_.foo.C2#m4()Lscala/Nothing;. => final def m4: Nothing
-      |_root_.foo.C3# => sealed class C3
-      |_root_.foo.C3#`<init>`()V. => primaryctor <init>: ()foo.C3
-      |_root_.foo.C3#m3()I. => def m3: Int
-      |_root_.foo.C3#toString()Ljava/lang/String;. => def toString: ()String
-      |_root_.foo.M. => final object M
-      |_root_.foo.M.C1# => case class C1
-      |_root_.foo.M.C1#`<init>`()V. => primaryctor <init>: ()foo.M.C1
-      |_root_.foo.M.C2# => class C2
-      |_root_.foo.M.C2#[T] => covariant typeparam T
-      |_root_.foo.M.C2#[U] => contravariant typeparam U
-      |_root_.foo.M.C2#`<init>`()V. => primaryctor <init>: ()foo.M.C2[T,U]
-      |_root_.foo.M.i1()Lscala/Nothing;. => implicit def i1: Nothing
-      |_root_.foo.M.l1. => lazy val l1: Nothing
-      |_root_.foo.T# => trait T
-      |_root_.foo.T#f1. => private val f1: Nothing
-      |_root_.foo.T#f2. => private val f2: Nothing
-      |_root_.foo.T#f3. => private val f3: Nothing
-      |_root_.foo.T#f4. => protected var f4_=: (x$1: Nothing)Unit
-      |_root_.foo.T#f5. => protected var f5_=: (x$1: Nothing)Unit
-      |_root_.foo.T#f6. => protected var f6_=: (x$1: Nothing)Unit
-      |_root_.foo.package. => packageobject package
-  """.trim.stripMargin
-  )
+  def targeted(markup: String, fn: m.Database => m.Name => Unit)(
+      implicit hack: OverloadHack2): Unit = {
+    test(markup) {
+      val (database, names) = computeDatabaseAndNamesFromMarkup(markup)
+      names match {
+        case List(name1) => fn(database)(name1)
+        case _ => sys.error(s"1 chevron expected, ${names.length} chevrons found")
+      }
+    }
+  }
 
-  sugars(
-    """
-      |package commandeer
-      |import scala.language.higherKinds
-      |trait CommandeerDSL[Host]
-      |object CommandeerDSL {
-      |  def apply[Host, DSL <: CommandeerDSL[Host]](host: Host)(implicit dsl: DSL): DSL = dsl
-      |}
-      |trait Foo
-      |object Foo {
-      |  implicit val fooDSL: FooDSL = new FooDSL {}
-      |}
-      |trait FooDSL extends CommandeerDSL[Foo]
-      |object RunMe {
-      |  CommandeerDSL(null.asInstanceOf[Foo])
-      |}
-  """.trim.stripMargin,
-    """
-      |[333..333) [commandeer.Foo, commandeer.FooDSL]
-      |[357..357) (commandeer.this.Foo.fooDSL)
-  """.trim.stripMargin
-  )
+  def targeted(markup: String, fn: m.Database => (m.Name, m.Name) => Unit)(
+      implicit hack: OverloadHack3): Unit = {
+    test(markup) {
+      val (database, names) = computeDatabaseAndNamesFromMarkup(markup)
+      names match {
+        case List(name1, name2) => fn(database)(name1, name2)
+        case _ => sys.error(s"2 chevrons expected, ${names.length} chevrons found")
+      }
+    }
+  }
 
-  sugars(
-    """
-      |class C[T]
-      |object C {
-      |  implicit def int: C[Int] = new C[Int]
-      |  implicit def list[T: C]: C[List[T]] = ???
-      |}
-      |
-      |class X
-      |object X {
-      |  implicit def cvt[T: C](x: T): X = ???
-      |}
-      |
-      |object M {
-      |  Nil.map(x => 2)
-      |
-      |  def c[T: C] = ???
-      |  M.c[List[Int]]
-      |
-      |  def x(x: X) = ???
-      |  x(42)
-      |
-      |  def i[T](t: T) = ???
-      |  i(new C[Int])
-      |}
-  """.trim.stripMargin,
-    """
-      |[191..191) [Int, List[Int]]
-      |[199..199) (immutable.this.List.canBuildFrom[Int])
-      |[237..237) (C.list[Int](C.int))
-      |[263..265) X.cvt[Int](*)(C.int)
-      |[294..294) [C[Int]]
-  """.trim.stripMargin
-  )
+  def targeted(markup: String, fn: m.Database => (m.Name, m.Name, m.Name) => Unit)(
+      implicit hack: OverloadHack4): Unit = {
+    test(markup) {
+      val (database, names) = computeDatabaseAndNamesFromMarkup(markup)
+      names match {
+        case List(name1, name2, name3) => fn(database)(name1, name2, name3)
+        case _ => sys.error(s"3 chevrons expected, ${names.length} chevrons found")
+      }
+    }
+  }
+
+  def targeted(markup: String, fn: m.Database => (m.Name, m.Name, m.Name, m.Name) => Unit)(
+      implicit hack: OverloadHack5): Unit = {
+    test(markup) {
+      val (database, names) = computeDatabaseAndNamesFromMarkup(markup)
+      names match {
+        case List(name1, name2, name3, name4) => fn(database)(name1, name2, name3, name4)
+        case _ => sys.error(s"4 chevrons expected, ${names.length} chevrons found")
+      }
+    }
+  }
 }
